Attempting to side-step error message when using shift
adamsmd reported an error when using `shift` and there were no arguments
to shift. I've seen this error when using an argument, such as with
`shift 2` but never have without an argument.

Moved the troublesome `shift` to within an `if` statement that will
check for the presence of additional arguments. This is my proposed
solution to #24.